Instructions
Adding the instructions for this lab, as well as the modifications that the instructions call for to the project. The instructions call to what I'm calling compact the qsys platform design project into a .qip file, which I've added to version control. A .qip file is apparently more efficient to synthesize. Ideally, all this would be done via batch tcl scripting.